first edition ready
